% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_dna_to_aa.R
\name{dna_codons_to_aa}
\alias{dna_codons_to_aa}
\title{Amino acid sequence from DNA sequence}
\usage{
dna_codons_to_aa(codons)
}
\arguments{
\item{n}{string}
}
\value{
aa string
}
\description{
Creates a sequence of amino acids from a given DNA sequence.
}
\examples{
dna_to_aa(AAAATGTCGACGATGCATAGCTAGCTGATGCTACG)

}
