% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_dna.R
\name{random_dna}
\alias{random_dna}
\title{Random DNA sequence}
\usage{
random_dna(n)
}
\arguments{
\item{l}{integer}
}
\value{
dna string
}
\description{
Creates a n long sequence of random nucleotides.
}
\examples{
random_dna(10)
}
