% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_rna.R
\name{random_rna}
\alias{random_rna}
\title{Creating a random RNA sequence of size l.}
\usage{
random_rna(l)
}
\arguments{
\item{l}{length og RNA seq}
}
\description{
Creating a random RNA sequence of size l.
}
